backend_api/requirements_main.txt
fastapi
uvicorn
python-multipart
face_recognition
numpy
opencv-python
pillow
